requirements.txt
python-telegram-bot==21.3
openai>=1.42.0
python-dotenv==1.0.0
pydantic==2.10.0
aiosqlite==0.20.0
httpx>=0.27.0
Pillow==10.4.0